chore: rename getAQuote to get-a-quote and eventTypes to event-types
